+++ include/ap1/pnc_sim/sim_node.hpp
@@ -15,10 +15,9 @@
 #define AP1_SIM_NODE_HPP
 
 #include <cmath>
+#include <rclcpp/node_options.hpp>
 #include <stdexcept>
-#include "geometry_msgs/msg/point.hpp"
 #include "rclcpp/rclcpp.hpp"
-#include "std_msgs/msg/float32.hpp"
 
 #include "ap1/pnc_sim/sim.hpp"
 #include "ap1_msgs/msg/entity_state.hpp"
@@ -31,6 +30,28 @@ using ap1_msgs::msg::FloatStamped;
 using ap1_msgs::msg::EntityStateArray;
 using ap1_msgs::msg::LaneBoundaries;
 
+template<typename T>
+T to_car_frame(
+  const T& p,
+  float car_x,
+  float car_y,
+  float car_yaw
+) {
+  T out;
+
+  float dx = p.x - car_x;
+  float dy = p.y - car_y;
+
+  float c = std::cos(car_yaw);
+  float s = std::sin(car_yaw);
+
+  out.x = c * dx + s * dy;
+  out.y = -s * dx + c * dy;
+  out.z = p.z;
+
+  return out;
+}
+
 namespace ap1::sim {
 class SimNode : public rclcpp::Node {
 public:
@@ -41,7 +62,8 @@ public:
 
   // Default Constructor
   SimNode(const ap1::sim::Sim &sim)
-      : Node("sim_node"), sim(sim) {
+      : Node("sim_node"), sim(sim) 
+  {
     // Subscriptions
     throttle_sub_ = this->create_subscription<FloatStamped>(
       "/ap1/control/motor_power", 1,
@@ -65,24 +87,45 @@ public:
     );
 
     // Publishers
-    lanes_pub_ = this->create_publisher<LaneBoundaries>("/mapping/lanes", 1);
-    entities_pub_ = this->create_publisher<EntityStateArray>("/mapping/entities", 1);
+    speed_pub_ = this->create_publisher<FloatStamped>("/ap1/actuation/speed", 1);
+    lanes_pub_ = this->create_publisher<LaneBoundaries>("/ap1/mapping/lanes", 1);
+    entities_pub_ = this->create_publisher<EntityStateArray>("/ap1/mapping/entities", 1);
     
-
+    // Log
     RCLCPP_INFO(this->get_logger(), "Simulation ROS Node Started.");
   }
 
+  /**
+   * @brief Publish the sim's details to the output topics.
+   * Transforms the absolute coordinate space of the sim into relative car coords.
+   */
   void publish() {
-    // Publish Car Speed
+    float car_x = sim.car.x;
+    float car_y = sim.car.y;
+    float car_yaw = sim.car.yaw;
+
+    // Speed
     FloatStamped speed_msg;
     speed_msg.value = this->sim.car.speed_mps;
     speed_pub_->publish(speed_msg);
 
-    // Publish entities
-    entities_pub_->publish(this->sim.entities);
+    // Entities
+    auto entities_msg = sim.entities; // copy
+    for (auto& e : entities_msg.entities) {
+      e = to_car_frame(e, car_x, car_y, car_yaw);
+      e.gamma -= car_yaw; // rotate orientation into car frame
+    }
+    entities_pub_->publish(entities_msg);
 
-    // Publish lanes
-    lanes_pub_->publish(this->sim.lane);
+    // Lane
+    auto lane_msg = sim.lane; // copy
+    for (auto& p : lane_msg.left) {
+      p = to_car_frame(p, car_x, car_y, car_yaw);
+    }
+    for (auto& p : lane_msg.right) {
+      p = to_car_frame(p, car_x, car_y, car_yaw);
+    }
+    lanes_pub_->publish(lane_msg);
   }
 private:
   const ap1::sim::Sim &sim;